+++ apps/gizual-app/src/primitives/date-picker/date-picker.tsx
@@ -1,12 +1,14 @@
 import { DatePickerInput } from "@mantine/dates";
 
+import { DATE_DISPLAY_FORMAT } from "@giz/utils/gizdate";
+
 type DatePickerProps = {} & React.ComponentProps<typeof DatePickerInput>;
 
 /**
  * Wrapper around the Mantine DatePickerInput component.
  * @param props - The props for the DatePicker component.
  */
-function DatePicker({ ...props }: DatePickerProps) {
+function DatePicker({ styles, ...props }: DatePickerProps) {
   return (
     <DatePickerInput
       styles={{
@@ -17,9 +19,10 @@ function DatePicker({ ...props }: DatePickerProps) {
           padding: "0 0.5rem",
           minWidth: 150,
         },
+        ...styles,
       }}
       {...props}
-      valueFormat="DD MMM YYYY"
+      valueFormat={DATE_DISPLAY_FORMAT}
     />
   );
 }