+++ apps/gizual-app/src/primitives/timeline/event-handler.ts
@@ -145,14 +145,12 @@ export class TimelineEventHandler {
     this._forwardEvent("mousemove", e);
     if (this.vm.isContextMenuOpen) return;
     if (this.vm.tooltip) {
-      const date = getStringDate(
-        estimateDayOnScale(
-          this.vm.timelineRenderStart,
-          this.vm.timelineRenderEnd,
-          this.vm.viewBox.width,
-          e.clientX + this.vm.currentTranslationX - this.parentBBox.left,
-        ),
-      );
+      const date = estimateDayOnScale(
+        this.vm.timelineRenderStart,
+        this.vm.timelineRenderEnd,
+        this.vm.viewBox.width,
+        e.clientX + this.vm.currentTranslationX - this.parentBBox.left,
+      ).toDisplayString();
 
       let tooltipContent = date;
       let hoveringId = -1;
@@ -171,7 +169,7 @@ export class TimelineEventHandler {
 
               return `┏━ ${getDateFromTimestamp(
                 commit.timestamp,
-              ).toDateTimeString()} ${author}\n┗━ ${commitMessage}`;
+              ).toDisplayString()} ${author}\n┗━ ${commitMessage}`;
             })
             .join("\n");
         }